Smallish naming and stylistic changes

+++ src/fdtransfer/fdtransferServer.cpp
@@ -17,38 +17,32 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <sys/syscall.h>
-#include <dirent.h>
 #include <errno.h>
 #include <fcntl.h>
 #include <signal.h>
 #include <time.h>
-#include <sched.h>
-#include <fstream>
 #include <unistd.h>
-#include <sys/ioctl.h>
-#include <asm/unistd.h>
-#include <stdbool.h>
+#include <sys/types.h>
+#include <sys/stat.h>
+#include <sys/sendfile.h>
+#include <sys/syscall.h>
 #include <sys/wait.h>
 
-#include "fdTransfer_shared_linux.h"
+#include "fdtransfer.h"
 #include "../jattach/psutil.h"
 
-
-#define TMP_KALLSYMS_PATH "/tmp/async-profiler-kallsyms.%d"
 
 class FdTransferServer {
   private:
     static int _server;
     static int _peer;
+    static int copyFile(const char* src_name, const char* dst_name, mode_t mode);
     static bool sendFd(int fd, struct fd_response *resp, size_t resp_size);
 
   public:
     static void closeServer() { close(_server); }
     static void closePeer() { close(_peer); }
-    static bool bindServer(struct sockaddr_un *sun, socklen_t addrlen, bool oneshot);
+    static bool bindServer(struct sockaddr_un *sun, socklen_t addrlen, int accept_timeout);
     static bool acceptPeer(int *peer_pid);
     static bool serveRequests(int peer_pid);
 };
@@ -56,34 +50,24 @@ class FdTransferServer {
 int FdTransferServer::_server;
 int FdTransferServer::_peer;
 
-bool FdTransferServer::bindServer(struct sockaddr_un *sun, socklen_t addrlen, bool oneshot) {
+bool FdTransferServer::bindServer(struct sockaddr_un *sun, socklen_t addrlen, int accept_timeout) {
     _server = socket(AF_UNIX, SOCK_SEQPACKET, 0);
     if (_server == -1) {
-        perror("FdTransfer socket():");
-        return false;
-    }
-
-    if (-1 == bind(_server, (const struct sockaddr*)sun, addrlen)) {
-        perror("FdTransfer bind()");
-        close(_server);
+        perror("FdTransfer socket()");
         return false;
     }
 
     // Arbitrary timeout, to prevent it from listening forever.
-    if (oneshot) {
-        const int accept_timeout_s = 10;
-        const struct timeval timeout = {
-            accept_timeout_s,
-            0,
-        };
-        if (-1 == setsockopt(_server, SOL_SOCKET, SO_RCVTIMEO, &timeout, sizeof(timeout))) {
+    if (accept_timeout > 0) {
+        const struct timeval timeout = {accept_timeout, 0};
+        if (setsockopt(_server, SOL_SOCKET, SO_RCVTIMEO, &timeout, sizeof(timeout)) < 0) {
             perror("FdTransfer setsockopt(SO_RCVTIMEO)");
             close(_server);
             return false;
         }
     }
 
-    if (-1 == listen(_server, 1)) {
+    if (bind(_server, (const struct sockaddr*)sun, addrlen) < 0 || listen(_server, 1) < 0) {
         perror("FdTransfer listen()");
         close(_server);
         return false;
@@ -109,7 +93,7 @@ bool FdTransferServer::acceptPeer(int *peer_pid) {
 
     if (*peer_pid != 0) {
         if (cred.pid != *peer_pid) {
-            fprintf(stderr, "unexpected connection from PID %d, expected from %d\n", cred.pid, *peer_pid);
+            fprintf(stderr, "Unexpected connection from PID %d, expected from %d\n", cred.pid, *peer_pid);
             close(_peer);
             return false;
         }
@@ -147,13 +131,9 @@ bool FdTransferServer::serveRequests(int peer_pid) {
             // Otherwise, verify the thread belongs to PID.
             if (peer_pid == 0 || syscall(__NR_tgkill, peer_pid, request->tid, 0) == 0) {
                 perf_fd = syscall(__NR_perf_event_open, &request->attr, request->tid, -1, -1, 0);
-                if (perf_fd == -1) {
-                    error = errno;
-                } else {
-                    error = 0;
-                }
+                error = perf_fd < 0 ? errno : 0;
             } else {
-                fprintf(stderr, "target has requested perf_event_open for TID %d which is not a thread of process %d\n", request->tid, peer_pid);
+                fprintf(stderr, "Target has requested perf_event_open for TID %d which is not a thread of process %d\n", request->tid, peer_pid);
                 error = ESRCH;
             }
 
@@ -172,37 +152,56 @@ bool FdTransferServer::serveRequests(int peer_pid) {
             // it's simpler to copy the file to a temporary location and pass the fd of it (compared to passing the
             // entire contents over the peer socket)
             char tmp_path[256];
-            snprintf(tmp_path, sizeof(tmp_path), TMP_KALLSYMS_PATH, getpid());
-            {
-                // TODO from Linux 3.15, we can use memfds here. Also, from 3.11 we can use O_TMPFILE.
-                std::ifstream src("/proc/kallsyms", std::ios::binary);
-                close(creat(tmp_path, S_IRUSR | S_IWUSR));  // create it with 0600 so others can't read.
-                std::ofstream dst(tmp_path, std::ios::binary);
-                dst << src.rdbuf();
+            snprintf(tmp_path, sizeof(tmp_path), "/tmp/async-profiler-kallsyms.%d", getpid());
+
+            int kallsyms_fd = -1;
+            int error = copyFile("/proc/kallsyms", tmp_path, 0600);
+            if (error == 0) {
+                kallsyms_fd = open(tmp_path, O_RDONLY);
+                if (kallsyms_fd == -1) {
+                    error = errno;
+                } else {
+                    unlink(tmp_path);
+                }
             }
 
             struct fd_response resp;
             resp.type = req->type;
-            int kallsyms_fd = open(tmp_path, O_RDONLY);
-            if (kallsyms_fd == -1) {
-                resp.error = errno;
-            } else {
-                unlink(tmp_path);
-                resp.error = 0;
-            }
-
+            resp.error = error;
             sendFd(kallsyms_fd, &resp, sizeof(resp));
             close(kallsyms_fd);
             break;
         }
 
         default:
-            fprintf(stderr, "unknown request type %u\n", req->type);
+            fprintf(stderr, "Unknown request type %u\n", req->type);
             break;
         }
     }
 
     return false;
+}
+
+int FdTransferServer::copyFile(const char* src_name, const char* dst_name, mode_t mode) {
+    int src = open(src_name, O_RDONLY);
+    if (src == -1) {
+        return errno;
+    }
+
+    int dst = creat(dst_name, mode);
+    if (dst == -1) {
+        int result = errno;
+        close(src);
+        return result;
+    }
+
+    while (sendfile(dst, src, NULL, 1048576) > 0) {
+        // Keep going
+    }
+
+    close(dst);
+    close(src);
+    return 0;
 }
 
 bool FdTransferServer::sendFd(int fd, struct fd_response *resp, size_t resp_size) {
@@ -241,7 +240,7 @@ bool FdTransferServer::sendFd(int fd, struct fd_response *resp, size_t resp_size
 
 static int single_pid_server(int pid) {
     // get its nspid prior to moving to its PID namespace.
-    int nspid = -1;
+    int nspid;
     uid_t _target_uid;
     gid_t _target_gid;
     if (get_process_info(pid, &_target_uid, &_target_gid, &nspid)) {
@@ -264,7 +263,7 @@ static int single_pid_server(int pid) {
         return 1;
     }
 
-    if (!FdTransferServer::bindServer(&sun, addrlen, true)) {
+    if (!FdTransferServer::bindServer(&sun, addrlen, 10)) {
         return 1;
     }
 
@@ -296,7 +295,7 @@ static int path_server(const char *path) {
         return 1;
     }
 
-    if (!FdTransferServer::bindServer(&sun, addrlen, false)) {
+    if (!FdTransferServer::bindServer(&sun, addrlen, 0)) {
         return 1;
     }
 
@@ -331,9 +330,9 @@ static int path_server(const char *path) {
     }
 }
 
-int main(int argc, const char *argv[]) {
+int main(int argc, const char** argv) {
     if (argc != 2) {
-        fprintf(stderr, "usage: %s [pid]/[path]\n", argv[0]);
+        fprintf(stderr, "Usage: %s <pid>|<path>\n", argv[0]);
         return 1;
     }
